Update cyclinguk_nearme.module
Uncommented out autocomplete.js ref in cyclinguk_nearme.module.  On prod since adding this reference in the cycle_travel_mapping module modal popups in drupal are not working, particularly when editing content/views, etc. I've commented out the reference to autocomplete.js in cycle_travel_mapping and uncommented the reference here. This is an attempt to make the modal pop ups work again.